requirements.txt
python-telegram-bot==13.5
requests==2.31.0
python-dotenv==1.0.0 